BJS2-97215 fixed after review

+++ src/main/java/faang/school/postservice/dto/cache/CacheUserDto.java
@@ -17,5 +17,5 @@ public class CacheUserDto {
     private Long id;
     private String username;
     @TimeToLive(unit = TimeUnit.DAYS)
-    private Long ttlDays;
+    private int ttlDays;
 }
